requirements.txt
python-dotenv
streamlit
fastapi
uvicorn
google-generativeai